backend/requirements.txt
Django==1.7
ipython
django_extensions
djangorestframework-jwt
djangorestframework
django-cors-headers